Answer question, new question works
All major functionality works! It has the weird quirk carried over
from the reactnd-chirper-app where it takes you back to the homepage
before the store has been loaded into the component so the question
pops up suddenly, trying to fix that for next time. Another quirk is
that there is only one entry for the questions reducer, becuase all
the logic is already contained in the API!