Users and questions actions and reducers functional
All store reducers and actions are functional and log correctly on the
console! Also, added loading bar to the login page! Next, to add
percentages to the question page component and then make the login
page more user friendly. Next, add tabs for the questions blurbs. Finally, we will work on the css styling for
the app.