All functionality done and styling done
Implementing the pie-chart was a design decision to see if there are
alternatives to just having horizontal loading bars for the polling. I
settled on pie charts with tooltips to show how many votes are in each
category: an annoying side effect is that if one option has zero
votes, a big 0% would show up with no ability to do tooltips, so I had
to create a workaround where the tooltip made that aside comment.
Another possible drawback is for narrow screens because the menu bars
will start overlapping. I will test it on a mobile device to see what
happens but no fix is in the pipeline.